requirements.txt
streamlit
python-dotenv
openai
llama-index
llama-index-core
llama-index-llms-openai
nltk
pypdf